+++ cdm_reader_mapper/common/json_dict.py
@@ -2,11 +2,7 @@
 
 from __future__ import annotations
 
-import ast
-import datetime
 import json
-import logging
-from copy import deepcopy
 
 from .getting_files import get_path
 
@@ -126,101 +122,3 @@ def combine_dicts(list_of_files, base=None):
             json_dict = combine_dicts(new_list_of_files)
         combined_dict = update_dict(combined_dict, json_dict)
     return combined_dict
-
-
-def _get_table_keys(table):
-    keys = []
-    for key, kdict in table.items():
-        if not isinstance(kdict, dict):
-            keys.append("~".join([key, kdict]))
-            continue
-        for k, v in kdict.items():
-            keys.append("~".join([key, k]))
-    return keys
-
-
-def get_table_keys(table):
-    """DOCUMENTATION."""
-    if table.get("_keys"):
-        _table = deepcopy(table)
-        _table.pop("_keys")
-        return _get_table_keys(_table)
-    else:
-        return list(table.keys())
-
-
-def eval_dict_items(item):
-    """DOCUMENTATION."""
-    try:
-        return ast.literal_eval(item)
-    except Exception:
-        return item
-
-
-def add_nested_keys(table, table_path):
-    """DOCUMENTATION."""
-    keys_path = table_path.with_suffix(".keys")
-    if keys_path.is_file():
-        with open(keys_path) as fileObj:
-            table_keys = json.load(fileObj)
-            table["_keys"] = {}
-            for x, y in table_keys.items():
-                key = eval_dict_items(x)
-                values = [eval_dict_items(k) for k in y]
-                table["_keys"][key] = values
-    return table
-
-
-def open_code_table(table_path):
-    """DOCUMENTATION."""
-    table = open_json_file(table_path)
-    # Add keys for nested code tables
-    table = add_nested_keys(table, table_path)
-    # Expand range keys
-    expand_integer_range_key(table)
-    return table
-
-
-def expand_integer_range_key(d):
-    """DOCUMENTATION."""
-    # Looping based on print_nested above
-    if isinstance(d, dict):
-        for k, v in list(d.items()):
-            if "range_key" in k[0:9]:
-                range_params = k[10:-1].split(",")
-                try:
-                    lower = int(range_params[0])
-                except Exception as e:
-                    logging.error(f"Lower bound parsing error in range key: {k}")
-                    logging.error("Error is:")
-                    logging.error(e)
-                    return
-                try:
-                    upper = int(range_params[1])
-                except Exception as e:
-                    if range_params[1] == "yyyy":
-                        upper = datetime.date.today().year
-                    else:
-                        logging.error(f"Upper bound parsing error in range key: {k}")
-                        logging.error("Error is:")
-                        logging.error(e)
-                        return
-                if len(range_params) > 2:
-                    try:
-                        step = int(range_params[2])
-                    except Exception as e:
-                        logging.error(f"Range step parsing error in range key: {k}")
-                        logging.error("Error is:")
-                        logging.error(e)
-                        return
-                else:
-                    step = 1
-                for i_range in range(lower, upper + 1, step):
-                    deep_copy_value = deepcopy(
-                        d[k]
-                    )  # Otherwiserepetitions are linked and act as one!
-                    d.update({str(i_range): deep_copy_value})
-                d.pop(k, None)
-            else:
-                for k, v in d.items():
-                    expand_integer_range_key(v)